Add generation for overlay via config / cli
The current API allows for overlays to be easily generated if using the
API. However, there is no clear way of generating these overlay figures
if using the config of CLI.

While redundant, this uses the three_view_frame together with some
bids2table functionality to grab and assert existence of overlay files.
Currently, this is only generating figures for the static
three_view_frame. Will want to generate something for videos too (either
slices or 4D or both). Likely want to take advantage of `slice_video`
and / or `three_view_video` for this.